add order books api and swagger doc

+++ models/account.go
@@ -7,9 +7,9 @@ import (
 // Account ...
 type Account struct {
 	gorm.Model
-	UserID     int64
+	UserID     uint
 	User       User
-	CurrencyID int64
+	CurrencyID uint
 	Currency   Currency
 	Balance    float64 `json:"balance" sql:"DECIMAL(32,16)"`
 	Locked     float64 `json:"locked" sql:"DECIMAL(32,16)"`
