+++ models/fund.go
@@ -8,9 +8,9 @@ type Fund struct {
 	Name            string
 	Symbol          string   // eg btc_usd eth_usd
 	RightCurrency   Currency `gorm:"foreignkey:RightCurrencyID"`
-	RightCurrencyID int64
+	RightCurrencyID uint
 	LeftCurrency    Currency `gorm:"foreignkey:LeftCurrencyID"`
-	LeftCurrencyID  int64
+	LeftCurrencyID  uint
 	LimitRate       float64 `json:"limit_rate" sql:"DECIMAL(32,16)"`
 	MarketRate      float64 `json:"market_rate" sql:"DECIMAL(32,16)"`
 }
